requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.3
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
